requirements.txt
asgiref==3.8.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
distlib==0.3.9
Django==5.2.5
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.11
filelock==3.18.0
idna==3.11
inflection==0.5.1
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
platformdirs==4.3.8
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.31.2
virtualenvwrapper-win==1.2.7
